+++ src/utils/enums/sidebar-status.enum.ts
@@ -1,4 +1,5 @@
 export enum SidebarStatus {
   SIGNED_OUT = 'signedOut',
   SIGNED_IN = 'signedIn',
+  LOADING = 'loading',
 }